refactor: centralize run step helper
