+++ live_trade/main_liveTrade.py
@@ -11,32 +11,7 @@ import logging
 import sys
 from pathlib import Path
 
-
-async def _run_step(step: str, script: Path, *args: str) -> None:
-    """Run a script as a subprocess and wait for it to finish.
-
-    If *script* points to a Python file, run it as a module so the project
-    root is on ``sys.path``. Otherwise fall back to executing the file
-    directly with the current Python interpreter.
-    """
-    if script.suffix == ".py":
-        try:
-            module = ".".join(
-                script.with_suffix("")
-                .resolve()
-                .relative_to(Path(__file__).resolve().parents[1])
-                .parts
-            )
-        except ValueError:
-            module = ".".join(script.with_suffix("").parts)
-        cmd = [sys.executable, "-m", module, *args]
-    else:
-        cmd = [sys.executable, str(script), *args]
-    logging.info("Running %s: %s", step, " ".join(cmd))
-    proc = await asyncio.create_subprocess_exec(*cmd)
-    await proc.communicate()
-    if proc.returncode:
-        raise RuntimeError(f"{step} failed with code {proc.returncode}")
+from scripts.common import _run_step
 
 
 def _load_config(path: Path) -> dict:
